[nrf fromlist] explicitly configure python to be used by GN
Use more portable way to find python 3 in the system and
pass the path to GN which uses python to run various helper
scripts during the build process.
The reason is that on different operating systems 'python'
name, which GN defaults to, may point to either 'python2' or
'python3' or it may not exist at all. On Windows, on the
other hand, python 3 is distributed as python.exe executable,
so 'python3' name also could not be used globally.

Signed-off-by: Damian Krolik <damian.krolik@nordicsemi.no>