[nrf noup] add workflow for building CHIP tools
Add a manually triggered workflow for building Android
CHIPTool packages for arm and arm64 architectures, and
CHIP Tool for Linux x64 and arm64.
The packages are uploaded both as github artifacts and
release packages (as the latter require a release tag).

The doxygen workflow pushes doxygen HTML files to another
branch which is undesired on the sdk-connectedhomeip fork
as it causes significant growth of the repository size.

Signed-off-by: Damian Krolik <damian.krolik@nordicsemi.no>
Signed-off-by: Kamil Kasperczyk <kamil.kasperczyk@nordicsemi.no>