[nrf noup] add workflow for building CHIP tools
Add a manually triggered workflow for building Android
CHIPTool packages for arm and arm64 architectures,
CHIP Tool for Linux x64 and arm64, and OTA provider for
Linux.
The packages are uploaded both as github artifacts and
release packages (as the latter require a release tag).

The doxygen workflow pushes doxygen HTML files to another
branch which is undesired on the sdk-connectedhomeip fork
as it causes significant growth of the repository size.

Signed-off-by: Damian Krolik <damian.krolik@nordicsemi.no>
Signed-off-by: Kamil Kasperczyk <kamil.kasperczyk@nordicsemi.no>

Remove Android Chip Tool from release workflow

Android Chip Tool is no longer maintained so it must be
removed from the release workflow.

Remove dependency to openssl from CHIP tool build for linux

Currently the chip-tool that is released as a part of
sdk-connectedhomeip has dependency to openssl.
The issue is that Ubuntu 20.04 and Ubuntu 22.04 use different version
of OpenSSL and version built under 20.04 does not work under
22.04 (without some hacks to install the old openssl version)
and vice-verse.

- Removed the dependency to the dynamic library (OpenSSL) to build
Chip-Tool with the mbedTLS.