[nrf noup] Use "continue-on-error" for examples in CI
Make building nRF Connect examples optional in github
workflows in a sense that they do not fail the workflow
if they do not compile. The reason is that the CI is often
run for downstream patches, that are required to proceed
with changes in NCS or Zephyr and we do not want to force
developers to update examples in sdk-connectedhomeip at
that moment - they will be updated in the upstream later.

Signed-off-by: Damian Krolik <damian.krolik@nordicsemi.no>