[nrf fromtree][nrfconnect] Fix for jsonschema documentation and factory data script
In the factory data script, there was no checking if a user has
jsonschema module installed. That caused an error even if the
user did not want to use that way of JSON validation.

Added checking a jsonschema module in python script for generating
factory data and added proper information in the documentation about
how to obtain jsonschema python module.