[nrf noup] Switch to statically allocated heap
* Use malloc/free replacements based on statically allocated
Zephyr's sys_heap to provide better control of RAM usage.

*  Fixed hard fault once advertising mdns records

In mDNS code there isn't a check that would verify if
memory was allocated successfully. In case heap is too small
the application will fall into hard fault due to usage of
non-allocated memory.

* Increased heap size for Wi-Fi.

Empirically it was measured that heap size needed to pass the
Wi-Fi commissioning is ~25k. To be on the safe side let's set it
to 28k.