[nrf noup] [nrfconnect] Introduce the WiFi connection recovery feature
The WiFi connection recovery feature allows device re-scanning
and re-connecting to the known WiFi network after the device's reboot
and when the known SSID has not been found during the last scan.
The connection recovery interval is doubled with
every occurrence to the defined maximum value and then its value
depends on the maximum value +- the defined random jitter.

After restoring the connection, the Connection Recovery
Interval is restored to the defined minimum value after elapsing
of the defined delay to avoid frequent reconnections to a
poor link quality network.

[nrfconnect] WiFi connection recovery refinements:
  * simplify recovery time resetting (due to connection status
    reporting limitations)
  * remove CHIP_WIFI_CONNECTION_RECOVERY_RESET_DELAY (not needed now)
  * avoid duplicated recovery timeout bumps
  * always start the recovery with kConnectionRecoveryMinIntervalMs
  * fixed some edge cases and memory leaks
  * minor cleanup
  * logging refinements
  * Kconfig description fixes after tech writer review
  * Move new Kconfig definitions to the Kconfig.features

Signed-off-by: Marcin Kajor <marcin.kajor@nordicsemi.no>