[nrf noup] Added script for downloading recommended zap package
The helper script is going to find recommended ZAP version in the
Matter repository, and download and extract the package from the
ZAP releases.

Signed-off-by: Kamil Kasperczyk <kamil.kasperczyk@nordicsemi.no>
Signed-off-by: Michał Szablowski <michal.szablowski@nordicsemi.no>

[nrf noup] Fixed output generated by get_zap.py on macOS

The zap executable location in script and printed path are not
correct.

[nrf noup] Fixed package names for zap binaries

The URLs for the ZAP tool binaries changed, so script for
downloading it had to be aligned.

Signed-off-by: Kamil Kasperczyk <kamil.kasperczyk@nordicsemi.no>

[nrf noup] Fix zap instalation for MacOs

Fix updating get_zap.py script, which ensure zap
will work on the MacOs systems.

Signed-off-by: Patryk Lipinski <patryk.lipinski@nordicsemi.no>
(cherry picked from commit c9cc3219269b049feadbe71b07d2796293ddd8ba)
Signed-off-by: Adrian Gielniewski <adrian.gielniewski@nordicsemi.no>