[nrf noup] Conditionally disable all unnecessary workflows
Currently, in sdk-connectedhomeip we don't need to run
upstream's workflows for all platforms because we use
sdk-nrf workflows to verify our PRs.

We decided to remove all of them except Lint Code Base and
Unit Tests from the 'pull request' and 'push' triggers and
leave the possibility to run them manually.

This commit cannot be pushed to the Matter upstream repository!