[nrf fromlist] Fix handling of certificate validity policy in CHIPDeviceControllerFactory.
When we created a new system state, we lost track of the certificate validity
policy (since it's not stored in the system state and was not copied from the
old one to the new one).

Just store it in the controller factory like other things of that sort, so we
make sure to propagate it to all system states.