[nrf noup] Add a workaround to WiFi recovery
During the recovery procedure, if the net_mgmt function returns
the value "-16" the device cannot recover the WiFi network
until the reboot is called.

This workaround allows the device to try to rescan the WiFi
network even if the net_mgmt function returned an error.
This workaround works only, when the WiFi recovery mechanism
is active.