[nrf fromtree] Fix cancellation of subscriptions to work correctly. (#28569)
* Fix cancellation of subscriptions to work correctly.

When a subscription came in with KeepSubscriptions set to false, we would just
directly delete the ReadHandlers for the subscriptions not being kept, instead
of calling Close().

That meant we skipped deleting subscription persistence data, and also failed to
correcly update the reporting engine's round-robin reporting state, which could
lead to subscriptions not being serviced fairly.

The fix is to just call Close() just like we do for out-of-resource eviction,
instead of manually deleting the object.

* Fix build issue.