[nrf noup] Optimize RAM usage
- Reduce number of network buffers and packets
This commit reduces the number of network buffers and packets used by
the nrfconnect platform.

- Add support for printing mbedTLS heap
This commit adds support for printing the peak usage of mbedTLS heap.

- Reduce number of Matter stack packet buffers
This commit reduced the number of available Matter's packet buffers to
8 which affects RAM usage.

- Introduce KConfig option to enable RAM profiling

This commit introduces the KConfig option that enables a set of
features useful for profiling the RAM memory.

- Disable SSL server support for Matter

This commit dissables unused support for SSL server.

- Reduce size of platform event queue

This commit overwrites the default size of platform event queue and
sets it to 64.

- Reduce mbedTLS heap and OpenThread stack sizes

This commit reduces the mbedTLs heap and OpenThread stack sizes when
Joiner role is enabled by not used (case of OT libraries).

- Reduce kernel heap size for Matter over Wi-Fi

This commit aligns the default heap size to the value used in the
Shell example.

- Optimize buffer usage of nRF700X driver for Matter

Matter device does not need to operate as high-performance STA, and so
this commit reduces the default number of buffers. As a result the heap
used by the supplicant may be significantly reduced.

- Added the missing conditions to HEAP configs.

Decrease stack sizes according to the peak results

Signed-off-by: Łukasz Duda <lukasz.duda@nordicsemi.no>
Signed-off-by: Arkadiusz Balys <arkadiusz.balys@nordicsemi.no>