[nrf noup] Remove recovery mechanism for Wi-Fi connection
Currently, the Wi-Fi driver supports the recovery mechanism and it
seems that it works properly so we can remove our recovery mechanism
when the connection is lost. Apart from losing the network,
we need to leave the recovery mechanism to work after the device's
reboot because we need to scan periodically to find a known network.