[nrf noup] Increase the maximum number of prefixes to match the maximum number of IPv6 addresses per interface.
Zephyr ipv6_nbr implementation requires the given address
to have a matching prefix set on the interface. Otherwise,
the default router is used for sending neighbor advertisement
and as a result, in case there are multiple routers in the network,
the packet can be sent to the invalid interface (not the one which
issued neighbor solicitation).

Signed-off-by: Marcin Kajor <marcin.kajor@nordicsemi.no>