[nrf fromlist] Clear endpoint bitmask before reporting the PartList attribute
When we set a low value to the minimum subscription interval for
subscription for the PartList attribute of the Description cluster,
and then remove a dynamic endpoint the related bitmask may change
after reporting it to a controller.
The controller receives the old number of endpoints and does not
update it within the following subscriptions until the next change
of the PartList attribute occurs.

To resolve the issue the endpoint bitmask should be cleared before
reporting the PartList attribute of the Descriptor cluster to avoid
race conditions.

[nrf] -> There is an old API and there is bitmask operation instead of
the method Clear!