[nrf noup] Add a workaround to the reporting engine
Currently, in the Matter 1.1.0.1 revision, there is a problem in
the reporting engine that causes wrong logic with marking mutated
change paths to send the report later. In the SetDirty method, we
called handler->setDirty before calling the InsertPathIntoDirtySet
method which caused the report not to be sent properly because
there was no new attribute to be sent on the DirtySet and the
generation version was forgotten, so the next data report did not
contain an update. It occurs when the controller sets too low value
for a minimal subscription interval and the device tries to send
a report scheduled to be sent immediately after the event occurs
without waiting to Inserting a path to the DirtySet.

This issue seems to be repaired in the Matter 1.2 revision, so we
can provide a workaround, and then revert/remove that PR during the
next synchronization with Matter SDK.