[nrf fromlist] [dns-sd] Untangle dependencies between Resolver and ResolverProxy
The relationship between Resolver and Resolver Proxy is
unclear as both classes depend on each other and Resolver
Proxy is implemented in a source file of concrete Resolver
implementation. This makes it impossible to build Matter
library with several Resolver implementations and select
one to be used at runtime.

The main reason for introducing Resolver Proxy was the need
for running multiple parallel DNS-SD queries. Address this
by adding Discovery Delegate parameter to the Resolver
interface methods for starting and stopping the discovery.

Then implement Resolver Proxy as a convenience class on top
of the Resolver interface. Additionally, simplify the
Commissionable Node Controller class by injecting a test
Resolver into Resolver Proxy.