[nrf noup][zephyr] Enabled support for bonding in the BLEManager
Currently the Zephyr BLE Manager rotates Bluetooth addresses
on every boot and it does this by creating new Bluetooth identity.
Because of that the Zephyr stack does not create default Bluetooth
ID, which is required e.g. for the bonding purposes.

Added creating two separate Bluetooth identities - for the Matter
service advertising and for the bonding purposes.

Signed-off-by: Kamil Kasperczyk <kamil.kasperczyk@nordicsemi.no>