[nrf fromlist] [app] Do not crash if CommandHandler fails to allocate packet
CommandHandler uses VerifyOrDie when adding a status to be
sent to the requestor. If the device runs out packet buffers
and CommandHandler fails to allocate a packet for the status,
the device crashes.

Triggering the crash requires many commands to arrive in the
device around the same time, which is probably rare but
possible.

Signed-off-by: Damian Krolik <damian.krolik@nordicsemi.no>