[nrf noup] Add custom west commands to ease ZAP usage
Add "zap-gui" west command that runs ZAP GUI to edit
the requested ZAP file.

Add "zap-generate" west command that generates the cluster
code out of the requested ZAP file.

Both commands verify if a correct version of ZAP package
is installed in .zap-install directory in the Matter SDK
module, and update the installed version accordingly.

The commands also try to find the ZAP file in the current
directory if the ZAP file path is not explicitly provided.