[nrf fromtree] [Zephyr] Clear ITS keys on factory reset
Added a kconfig option to enable clearing keys stored in the
PSA Internal Trusted Storage. If the kconfig is selected then
during the factory reset process all keys stored in the PSA
ITS will be destroyed.

This mechanism ensures that all PSA ITS crypto materials that
belong to the Matter stack will be removed.