[nrf noup] Shedule factory reset in Server using Matter Shell command
It is better to call the ScheduleFactoryReset method from
the Server, because it also removes all fabrics and emits the Leave
event whereas calling InitiateFactoryReset() removes only the
persistent storage entries.

In this way, the Matter Shell command works in the same way as
other factory reset sources.

We cannot upmerge it in an easy way because many vendors do not
use chip server and there are multiple building issues.

Signed-off-by: Arkadiusz Balys <arkadiusz.balys@nordicsemi.no>