[nrf toup] Include EXTRAVERSION in Matter software version string
When a Matter application provides the VERSION file, the
Matter software version string is set to 4-number string,
such as "2.7.99+0" and the EXTRAVERSION field, which is
typically used to denote the branch or tag name, such as
"dev" or "rc1", is ignored.

If the EXTRAVERSION field is non-empty, set the Matter
software version string to a string such as "2.7.99-dev+0".

(cherry picked from commit 731aac5c8941bf0f0a9f10c759f533e5b01f80eb)
Signed-off-by: Adrian Gielniewski <adrian.gielniewski@nordicsemi.no>