[nrf fromtree] [ESP32]: increment the total operational hours inline without log (#37058) (#37069)
* [ESP32]: increment the total operational hours inline without log

Since total-operational-hours is a critical information, it
intentionally uses the FreeRTOS timer to increment the values, so that
it should not be a victim of PostEvent failures.

Earlier it used to call WriteConfigValues which has the ChipLogProgress
and logging from the timer stack may overflow the timer stack. So,
inlined the implementation which do not log.

* Restyled by clang-format

* add reasoning behind the change as comment

---------

Co-authored-by: Restyled.io <commits@restyled.io>
(cherry picked from commit c05637672dfad61be1054b8c229d2a585ac3b97c)
Signed-off-by: Adrian Gielniewski <adrian.gielniewski@nordicsemi.no>