[nrf noup] Workaround for the KMU usage with the new NOC approach
Now, we copy the NOC key from volatile to persistent space
during the commissioning OP keypair.

Currently, psa_copy_key does not work for KMU, so we must
export this key and import it again as a workaround.

This PR is a workaround and can be fully reverted once the
NCSDK-34997 issue is resolved.

Signed-off-by: Arkadiusz Balys <arkadiusz.balys@nordicsemi.no>

Restyled by clang-format