[nrf toup][nrfconnect] Increase ICD observer pool size
The ICDMnager observer for the DNS-SD server was silently
not registered which causes the OnICDModeChange function
to be called on change from SIT to LIT.

The root cause is that in Matter 1.5, there is
an additional observer, and we have a too small pool size to
fit the 3rd one for the DNS-SD Server.

Increase the CHIP_CONFIG_ICD_OBSERVERS_POOL_SIZE to 3 for
the Nordic platform.

Signed-off-by: Arkadiusz Balys <arkadiusz.balys@nordicsemi.no>